build: add sanitizers
Add optional sanitizers to the build system, for GCC and Clang. I don't
know what are the exact flags for MSVC.

The output from sanitizers will be more clear with debug info, so don't
force the Release build type and just set it as default instead.

For example, to build the program with UBSAN and ASAN enabled:

    cmake <build dir> -DCMAKE_BUILD_TYPE=Debug -DENABLE_UBSAN=ON -DENABLE_ASAN=ON

Configure sanitizers to log their output to files instead of stderr:

    export UBSAN_OPTIONS='log_path=ubsan,print_stacktrace=1'
    export ASAN_OPTIONS='log_path=asan'
    export TSAN_OPTIONS='log_path=tsan'

And then run the program.

More information about the sanitizers:

https://clang.llvm.org/docs/UndefinedBehaviorSanitizer.html
https://clang.llvm.org/docs/AddressSanitizer.html
https://clang.llvm.org/docs/ThreadSanitizer.html