perf: remove std::function in ItemSorter::sort
std::function can store any type of function, including for example
functions that have state. Because compilers can't figure out at compile
time what type of function is it going to be or how much memory is going
to be needed to store the state, std::function resorts to polymorphism,
which can have quite big runtime cost.

std::function can be a good thing if you want to provide a customization
point in a library, but it should be avoided in high performance code.

I can't tell how much overall performance impact it has here, because
there are no benchmarks to measure it, but I ran some micro benchmarks
with just std::sort. Using normal function pointers on my machine is
1.5x faster than the baseline std::function solution:

    std::sort(begin, end, cond ? func1 : func2);

And pulling it out to separate branches (like in the PR) is 2x faster:

    if (cond) {
        std::sort(begin, end, func1);
    } else {
        std::sort(begin, end, func2);
    }

Additionally, compare functions are now marked as static, so the
compiler knows that they are visible only to this file, which can
potentially enable more optimizations like function inlining.