refactor: use static
Generally everything that is declared and visible only in a single
translation unit (.cpp file) should be marked as static. First of all,
this avoids linker errors (one definition rule violations) if you have
symbols with the same name defined in multiple files. Second, it enables
more optimizations. The compiler is only aware of what's happening in a
single file at the time, and static gives it a hint that it can be
optimized in the context of that particular file.