refactor: C header includes
Use the C++ versions of C headers and wrap other system headers in
extern "C". During compilation symbol names in C++ are "mangled" to
include type information. C doesn't do this, so if you have a
declaration in both C and C++ that seems to look exactly the same, C++
might not connect the dots that this is supposed to be the same
function. This isn't usually a problem with libc and system headers, but
still it's more correct to wrap the declarations of C functions in
extern "C" block. C++ versions of C headers (stdio.h -> cstdio) do this
for you.